+++ packages/miew/src/io/parsers/MMTFParser.js
@@ -463,7 +463,7 @@ class MMTFParser extends Parser {
     this._joinSynonymousChains();
 
     this._complex.finalize({
-      needAutoBonding: false,
+      needAutoBonding: this.resolveAutoBonding(false),
       detectAromaticLoops: this.settings.now.aromatic,
       enableEditing: this.settings.now.editing,
       serialAtomMap: this._serialAtomMap,