+++ packages/miew/src/io/parsers/MOL2Parser.js
@@ -276,7 +276,7 @@ class MOL2Parser extends Parser {
     this._finalizeMolecules();
 
     this._complex.finalize({
-      needAutoBonding: false,
+      needAutoBonding: this.resolveAutoBonding(false),
       detectAromaticLoops: this.settings.now.aromatic,
       enableEditing: this.settings.now.editing,
       serialAtomMap: this._serialAtomMap,